requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.11
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
